Various refactorings.
1) create procUtil.js for process utilities
2) split test files by test suite
3) add pdfjs to octane regression!
